app/pyproject.toml
[project]
name = "voxvibe"
version = "0.4.2"
description = "A voice dictation application for Linux that captures audio and transcribes it to text using Whisper"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.1.1",
    "litellm>=1.74.0",
    "mistralai>=1.0.0",
    "numpy>=2.3.0",
    "pynput>=1.7.6",
    "pyqt6>=6.9.1",
    "qt-material>=2.17",
    "sounddevice>=0.5.2",
    "soundfile>=0.12.1",
    "speechmatics-batch>=0.4.1",
]

[project.scripts]
voxvibe = "voxvibe.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.11.13",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["I", "F"]
